+++ src/app/shared/fhir-backend/fhir-backend.service.spec.ts
@@ -334,6 +334,7 @@ describe('FhirBackendService', () => {
                 cacheName: null,
                 retryCount: false,
                 cacheErrors: true,
+                cacheAbort: false,
                 priority: PRIORITIES.NORMAL
               }
             );
@@ -358,6 +359,7 @@ describe('FhirBackendService', () => {
               cacheName: null,
               retryCount: false,
               cacheErrors: true,
+              cacheAbort: false,
               priority: PRIORITIES.NORMAL
             }
           );